feat: add batchDenoms to take success tracking event
